Wallet: allow to create "deep-locked" Wallet
A "deep-locked" wallet is a byte[], consisting of
- The private key, encrypted using the wallet pin and a hard KDF
- The salt used for the KDF
- A 4-byte "checksum" (the first four byte of the hash of the public key)

The public key and address corresponding to a deep-locked wallet are not available while such a wallet is locked. It is thus required to know the wallet pin to import a deep-locked wallet, even if the private key portion is not used.