Use packaging addons when building for Travis
Travis sometimes builds on a more recent CPU than the one used to run
the tests leading to illegal instruction errors when running the tests.

With this commit, the addon will be built for several different CPU
optimizations and the correct one will be loaded at runtime, so that the
build CPU optimizations don't need to match the test CPU optimizations.