Ensure expectMessage callback is always invoked
expectMessage() is supposed to invoke a callback when an expected message is
not returned before the specified timeout. However the corresponding timer is
removed when the related connection is closed, leading to the callback never
being invoked.

This will invoke all callbacks of expected messages when a connection is closed.