Temporarily move node-deb to production dependencies
node-deb must be moved to the production dependencies section in order
to use it to produce packages without any dev-dependencies (otherwise it
wouldn't be available to build the packages since that's done after the
removal of the dev-dependencies).

Eventually we should implement the proper fix (which means replacing
node-deb with standard Debian tools to create packages).